[DOMD] xen: Implement PCI specific checks for hardware domain
On x86 it is assumed that all the PCI host bridges live in a hardware
domain and only Domain-0 is now supported. On ARM it is possible that
real PCI host bridge hardware is passed through to a driver domain or
other entity in the system, thus a simple is_hardware_domain check is
not enough.

Introduce a PCI dedicated pair of functions to check whether bridge's
segment/bus physically owned by a domain given and means to get the owner
domain of the bridge.

Assume for now that all host bridges are owned by a single driver
domain.

Signed-off-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>