xen/arm: Intercept vPCI config accesses and forward them to emulator
This is needed for supporting virtio-pci.

In case when the PCI Host bridge is emulated outside of Xen
(IOREQ server), we need some mechanism to intercept config space
accesses on Xen on Arm, and forward them to the emulator
(for example, virtio backend) via IOREQ request.

Unlike x86, on Arm these accesses are MMIO, there is no CFC/CF8 method
to detect which PCI device is targeted.

Please note, this patch only covers a case when both PCI Host bridge
and all PCI devices behind it are located in the same emulator (i.e. Qemu).
For a case when PCI devices are located in different emulators we would
likely need to emulate dummy PCI Host bringe here in Xen for the integration.

TODO: Split into several patches. Also сheck whether it won't break
the PCI passthrough as we might have both features enabled at
the same time.

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>