libxl: add support for emulated TPM on ARM
QEMU gained support for running as IOREQ server to do device emulation
for ARM guests. It is done using a "xenpv" emulated machine for aarch64
targets (same name as the x86 "xenpv" machine for PV backends.) It
supports both PV backends and IOREQs.

Introduce a new xl config option "tpm" to enable TPM emulation via QEMU.
We are not reusing the existing vtpm option because that is tied to a PV
frontend and backend protocol that has nothing to do with the TPM
emulation provided by QEMU. vTPM is configured via xenstore, while
emulated TPM via the QEMU command line.

If tpm=1 pass the right command line parameters to QEMU so that it can
connect to swtpm and provide TPM emulation for Xen guests. Also if tpm=1
request xenpv QEMU to be spawned.

Signed-off-by: Stefano Stabellini <stefano.stabellini@xilinx.com>