pvh dom0: fix MSI-x initialization for passed-through devices
As part of XSA-337, reading BARs is not allowed once the device is
assigned to a guest. The reading of BARs is performed earlier by Dom0
in the pcistub probe. For this, Dom0 needs to support the hypercalls
PHYSDEVOP_prepare_msix and PHYSDEVOP_release_msix. As of now, these
physdevop commands are available only to pv guests.
Enable them also for hvm hardware domains to fix MSI-x initialization
when Dom0 is pvh.

Signed-off-by: Xenia Ragiadakou <xenia.ragiadakou@amd.com>