arm: new VGIC: its: Introduce ITS device list
Add the vgic_its_device structure and the list of devices to the vgic_its.
Implement the functions to allocate, free, add and delete the device from
the list. Add the function to find the device in the list by the device id.
And make them avaliable to the HW ITS code.

Signed-off-by: Mykyta Poturai <mykyta_poturai@epam.com>