tools: Allow for automatic selection of virtio pci devid numbers
Add calls to pci_devid_pool dmctl to allow for automatic selection of
virtio pci devid numbers. This is done in two passes to ensure that
explicitly set devids are taken first.

Later in the process, we also replace the XL_PCI_SBDF_DEV_x strings in
the dm_args array with the actual devids.

Signed-off-by: Mykyta Poturai <mykyta_poturai@epam.com>